requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
